+++ src/main/java/nextstep/subway/application/exception/exception/AlreadyDataException.java
@@ -1,4 +1,4 @@
-package nextstep.subway.common.exception;
+package nextstep.subway.application.exception.exception;
 
 public class AlreadyDataException extends RuntimeException {
