+++ src/main/java/nextstep/subway/application/exception/exception/NotFoundDataException.java
@@ -1,4 +1,4 @@
-package nextstep.subway.common.exception;
+package nextstep.subway.application.exception.exception;
 
 public class NotFoundDataException extends RuntimeException {
