+++ src/main/java/nextstep/subway/application/exception/type/AlreadyExceptionType.java
@@ -1,4 +1,4 @@
-package nextstep.subway.common.type;
+package nextstep.subway.application.exception.type;
 
 public enum AlreadyExceptionType {
     ALREADY_LINE_STATION("이미 존재 하는 LineStation 이에요");
