+++ src/main/java/nextstep/subway/domain/Line.java
@@ -1,10 +1,15 @@
 package nextstep.subway.domain;
 
 
+import nextstep.subway.application.exception.exception.NotValidDataException;
 import org.springframework.util.StringUtils;
 
 import javax.persistence.*;
 import java.util.List;
+import java.util.Objects;
+
+import static nextstep.subway.application.exception.type.ValidExceptionType.NOT_VALID_COLOR;
+import static nextstep.subway.application.exception.type.ValidExceptionType.NOT_VALID_NAME;
 
 
 @Entity
@@ -25,9 +30,24 @@ public class Line extends BaseEntity {
     }
 
     private Line(String name, String color, Station upStation, Station downStation, int distance) {
+        checkValidNamePresent(name);
+        checkValidColorPresent(color);
+
         this.name = name;
         this.color = color;
         addLineStation(LineStation.of(upStation, downStation, distance));
+    }
+
+    private void checkValidColorPresent(String color) {
+        if (Objects.isNull(color)) {
+            throw new NotValidDataException(NOT_VALID_COLOR.getMessage());
+        }
+    }
+
+    private void checkValidNamePresent(String name) {
+        if (Objects.isNull(name)) {
+            throw new NotValidDataException(NOT_VALID_NAME.getMessage());
+        }
     }
 
     public void addLineStation(LineStation station) {
