+++ src/test/java/nextstep/subway/exception/AcceptanceTest.java
@@ -1,4 +1,4 @@
-package nextstep.subway.common;
+package nextstep.subway.exception;
 
 import io.restassured.RestAssured;
 import org.junit.jupiter.api.BeforeEach;
