+++ src/test/java/nextstep/subway/exception/DataBaseCleanUp.java
@@ -1,4 +1,4 @@
-package nextstep.subway.common;
+package nextstep.subway.exception;
 
 import org.springframework.beans.factory.InitializingBean;
 import org.springframework.beans.factory.annotation.Autowired;
