+++ src/test/java/nextstep/subway/lineStation/LineStationAcceptanceTest.java
@@ -2,7 +2,7 @@ package nextstep.subway.lineStation;
 
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
-import nextstep.subway.common.AcceptanceTest;
+import nextstep.subway.exception.AcceptanceTest;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
